+++ kwcoco/toydata.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-from .demo.toydata import *  # NOQA
-import warnings
-warnings.warn('kwcoco.toydata is deprecated, use kwcoco.demo.toydata instead',
-              DeprecationWarning)